Adding a new skybox
Do not think it will be the final skybox as it would be better to use for a sphere skybox rather than a cube

Added a new skybox, bottom of the skybox tranlation needs to be fixed. Also quessing the left part of the skybox is missing aswell it would be where the large checker box would of been. Made the Skybox smaller because 1000 is quite large 500 might be to large aswell could not see the skybox when far away from it.

Commeted out some unnecessary code in initalize models and cubes.

Have also removed the branch I was working on to get the game level model to work. Soultion that I thought might fix the warping problem was not going to work.

Added in a tree model to see if warping problem was still there will remove tree for pipeline soon.